Move conftest to root test directory
